M4 step 1: gbx-prng (binary-exact PRNG), audited migration, .rsav v2
Open the M4 build order's step 1 (oracle-rig.md §1/D-OR1). Replace the
splitmix64 placeholder with CotAB's recovered Turbo Pascal LCG and route
all game randomness through the one crate.

New crate `crates/gbx-prng` (pure, wasm-clean, serde-only leaf):
- `next()` = `state*0x08088405 + 1`; `random(n)` = exclusive `hi16 mod n`,
  ALWAYS advancing the state including `n == 0` (draws then returns 0 — no
  short-circuit, the D-OR1(b) draw-parity contract). `state`/`set_state`
  for .rsav and the oracle poke. `random_real()` deferred to M5.
- CI-safe tests incl. the draw-always `random(0)` test and LCG/`random`
  pins computed externally. Local-tier test re-derives the `[0xa55a,0xa5ee)`
  hash from the user's own binary and loud-fails on mismatch.
- Added to the CI wasm job's crate list.

Audited per-site migration (oracle-rig.md §6 ledger, one row per site):
- `EngineServices::roll`: CONFIRMED off-by-one fixed — `roll_uniform(max)`
  was inclusive `0..=max` and could return `operand+1`; now `random(max)`
  (exclusive). The mechanical rename `random(max+1)` would have frozen the
  bug. Same fix in the CLI host, which also carried a second RNG
  (xorshift64*) — removed, now draws from gbx-prng.
- `roll_dice`/`roll_die`/`EngineRoller`: `1 + random(size)`; `size == 0`
  now draws (was a non-drawing short-circuit / underflow panic).
- `VmRng` trait method `roll_uniform(inclusive_max)` -> `random(n)`.
- Fade dither (`apply_recolor_dithered`) moved OFF the PRNG to a
  deterministic position hash; signature drops `&mut dyn VmRng` (D-OR1(c)).

.rsav consequence (atomic in this commit): `SAVE_FORMAT_VERSION` 1 -> 2
(postcard would silently misload the old u64 state into the new u32 field);
golden recomputed by running the test. Engine seed narrowed to u32 at the
API surface; `ContainerHeader.seed` stays u64 (provenance, zero-extended).

Docs: SOURCES.md row; docket FD-26/FD-28 updated, FD-29 (roll_dice byte
truncation) and FD-30 (creation-roll draw order) filed; PLAN.md M4 H3
annotated.

+++ crates/gbx-engine/src/engine.rs
@@ -181,7 +181,7 @@ impl Engine {
     /// Boots from real `GameData`: the D-UI1/`boot.rs` asset slice, this
     /// session's hardcoded resident GEO block, and the initial resident ECL
     /// block (`ECL{GAME_AREA}.DAX` block [`INITIAL_ECL_BLOCK`]).
-    pub fn new(data: GameData, seed: u64) -> Result<Self, BootError> {
+    pub fn new(data: GameData, seed: u32) -> Result<Self, BootError> {
         let assets = boot::boot(&data)?;
         let geo_bytes = data.block(DEFAULT_GEO_FILE, DEFAULT_GEO_BLOCK)?;
         let geo = GeoBlock::parse(&geo_bytes)?;
@@ -209,7 +209,7 @@ impl Engine {
         symbol_sets: SymbolSets,
         geo: GeoBlock,
         data: GameData,
-        seed: u64,
+        seed: u32,
     ) -> Self {
         Self::build(font, symbol_sets, dummy_sky(), geo, data, seed)
     }
@@ -220,7 +220,7 @@ impl Engine {
         sky: [ImageBlock; 3],
         geo: GeoBlock,
         data: GameData,
-        seed: u64,
+        seed: u32,
     ) -> Self {
         let mut fb = Framebuffer::new();
         crate::frames::draw8x8_03(&mut fb, &symbol_sets)
@@ -260,7 +260,10 @@ impl Engine {
             sounds: Vec::new(),
             serial: 0,
             last_hash: None,
-            boot_seed: seed,
+            // Provenance header field stays u64 (D-OR1): the u32 live seed is
+            // zero-extended so `ContainerHeader.seed` and its byte layout don't
+            // churn. Resume never reads this — it restores `SaveState.prng`.
+            boot_seed: seed as u64,
             tick_count: 0,
             symbol_sets,
             sky,
